allow pure build of zig derivation
Prior to this commit, if nixpkgs was imported with a definite system, it
would *mostly* work as intended and not make impure calls to Nix's
builtins.CurrentSystem, except for here in zig_raw.nix, which doesn't
respect sources and pkgs. The solution is simple: just inherit sources
and pkgs when we make a versioned zig_raw in zig.nix.